+++ utils/__init__.py
@@ -1,8 +1,7 @@
 from .decrement import decrement
-from .errors import NotImplementedYetError
+from .errors import NotImplementedYetError, todo, unimplemented
 from .identity import identity
 from .increment import increment
-from .macros import todo, unimplemented
 from .min_max import min_max
 from .round import round
